Updated openstack/openstack Project: openstack/horizon  147a4ec70d970c91f16247e4caad9d0e458f70d9
Guard against instance row being displayed slowly in i9n tests

Sometimes due to slow Nova response in integration tests an instance
row is being checked for the presence of a specific status before a
row even appears.  Wait not only for the status appearance, but for
the row appearance as well.

Change-Id: Ifc58ef55a1e85c99ed442858c4f51a045603dc62
Closes-Bug: #1568387