Updated openstack/openstack Project: openstack-infra/subunit2sql  6a00ad755fba6cfe98d71357ff08173c0dd7e60c
Add stdout to indicate expiring runs operations

This commit adds 2 stdout messages to indicate to users that the
expiration is actually happening. We're seeing the expire cron job
on the infra db return 0 but the results are still in the db.
This is an attempt to provide some indication that things are
actually running.

Change-Id: I1278960110ff75a8e7e3df1572e9855a0c40e405