Updated openstack/openstack Project: openstack/keystone  614a6ef4c9b66e5fd2814ddf84e51d7a02683a6b
Fixes bug where the updated federated display_name is not returned

When shadowing a federated user, if the display name is changed, it
should get updated and returned in the user name attribute.  This patch
fixes a bug where the display_name was getting updated, but not the old
display_name was being returned.

Closes-Bug: #1566494

Change-Id: I155d3a9e4c90a3d22d0b30e35276c9ddbb65ae6d