Updated openstack/openstack Project: openstack/cinder  113beaf29891d72a89ddc7464656098b000dcf96
Fix driver.failover_host call in manager.py

We have things set on the driver to take a named arg here,
but were just passing in a positional value.

Fix that.

While we're at it, the HP and EMC drivers went their own way and used their own
parameters for failover.  We fix those up to match the other drivers including
the reference in driver.py.

Change-Id: I58e3d28e3185c2448556f75a21deecea5326a26f
Closes-Bug: #1564200