Updated openstack/openstack Project: openstack/cinder  69543b535c7168f4d2cf907e1bfe44006208a5db
Update SolidFire driver to store cluster info

Currently we have the ability to send commands to unconfigured SolidFire
clusters using the endpoint argument in our api calls.  That's not bad
but it has lead to some cruft and confusing code where it becomes
difficult to determine "which" endpoint we actually want, and if we
were able to pass it through the sometimes long chain of commands.

This change is a move to simplify things a bit.  Rather than pass
and endpoint variable around, this introduces the strategy of
setting an "active_cluster" in the driver.  This way an operation
can adjust this as needed based on the call being issued.

This also let's us consolidate all of the cluster specific
parameters in to a single data structure which in and of
itself is worthwhile.

Change-Id: Icdfb8c2c0e760a255db9387bf9df28616f2497aa