Updated openstack/openstack Project: openstack/os-net-config  2af4ee4b873b30d56c24cf1dfcad7bfabfedab37
Normalize operstate value for interfaces

The previous check for an interface operstate only looked for the
literal "UP" in caps, but in my environments I'm seeing operstate
returned as "up" in lower-case, which causes the _is_active_nic
check to fail incorrectly.  Example:

[root@overcloud-controller-0 heat-admin]# cat /sys/class/net/eth0/operstate
up

In this environment os-net-config is failing with the exception
in the linked bug.

Change-Id: I85c2d074ce43673c691523ca146ff7cdfdf1c7ca
Closes-Bug: 1569403