Updated openstack/openstack Project: openstack/governance  a21f11b6eb2c45ae3dbdf642016637e6b223d52a
Select Thierry Carrez as TC chair for Newton

The chair is responsible for making sure meetings are held
according to the rules described in our charter, and for
communicating the decisions taken during those meetings to the
Board of Directors and the OpenStack community at large.

If you'll have me, I'll be happy to serve again this cycle.
If you'd like to submit your own candidacy, please propose
an alternate patch. We'll vote on this at the next meeting.

Change-Id: I23f8953f510b2ec8bf4ab9def49891e2ebc2cc1e