Updated openstack/openstack Project: openstack-infra/project-config  7b09330da042517652748935bc5452f9f9cc7e0c
publish puppet-aodh in tarball for master

This is an experimental thing for now, to get some feedback.
The idea is to follow other Python projects and publish a release of
Puppet Modules on tarballs server.

This first implementation will publish a tarball of puppet-aodh module
for each patch sent in master.

Change-Id: Iac61083ed3d0a587186dfe0ba6f61ae7f93a746a