Updated openstack/openstack Project: openstack/neutron  153f535d55e0e2b99a1220605730470ae497da45
Refactor IPAM ip allocation method

After taking some time going through this method to figure out what it
is really doing, I found it could be written with less tangled logic
and maybe I could save time for the next person to come along and try
to figure it out.

Also, I'm likely to do some more in this area as part of the routed
networks work and this will help.

Change-Id: I4d96ea8953fb811ec26c61d91ec9989e2b7a3216