Updated openstack/openstack Project: openstack-infra/shade  6a89c4f3a111f6895074770b3b1913c6eb56b926
Consume config values for NAT destination

In a cloud that has multiple private networks and that requires
floating IPs for public access, it's not possible to know which private
network the floating IP should NAT to without the user providing
some configuration. os-client-config added support for expressing
such a thing in 1.17.0, so start consuming that.

Change-Id: Iee634bcdc933a4c8d793ab9b25b27d9ecd77cdcc