Updated openstack/openstack Project: openstack/ironic  ca69aa83f695aecd02d8d233bba4c0ea7504a679
deployment vmedia ops should not be run when not deploying

When a conductor fail-over occurs, the conductor _do_takeover method
calls task.driver.deploy.prepare(). In some drivers, this results in
task.driver.boot.prepare_ramdisk() being called which when dealing
with virtual media, is inappropriate to do when not performing
deployment operations as this can result in the deployment virtual
media being attached to a running host.

Change-Id: Ief8dc385ccb67432eed05813ca266dd1e316575c
Closes-Bug: #1565104