Updated openstack/openstack Project: openstack-infra/system-config  8d5b262e7e7d5203a159d1b6437cc5ae460443a3
Put uri_rewrite and data_dir in variables

We need these values in both the o_p:npm_mirror class and in the
o_p::mirror_update class. This indicates that we didn't use containment
as well as we should have, but that's fine and we can clean this up
later.

Change-Id: I1cbf5b30b5f2fa22d576cd23da9eb34942f6df76