Updated openstack/openstack Project: openstack/nova  3c522e8bf2dbf9ba3ca1261eaabc64a4f50d38f5
DB API changes for the nova-manage quota_usage_refresh command

This is the first patch in a sequence to implement the
blueprint.

The patch implements the database API changes that will be
needed by the new nova-manage quota_usage_refresh command.

A new API is added to refresh the quota usages of the
selected resources.

Another API is added to get the PER_PROJECT_QUOTAS, which
will be needed later in the patch sequence.

Implements blueprint: refresh-quotas-usage

Change-Id: Id36328b2f2016eab2d7194b88d3eee8d016786d1