Updated openstack/openstack Project: openstack/nova  74755c4ddbfa59c8b793317964303ba0244452f9
VMware: enable a resize of instance with no root disk

This addresses the issue when the user wishes to update the CPU
or memory of the instance.

No changes to disks will be done if there is no root disk.

Examples of when there is no root disk:
1. When an instance is booted from a cinder volume.
2. When booting from an ISO and the root disk size is zero

In these cases there will not be a disk created on the vmware
shared datastore.

Change-Id: I848fd3e57ec2f9cbc4fc31fd9ec3fe515fee4f32
Closes-bug: #1444446