Updated openstack/openstack Project: openstack-infra/shade  f944c6e3dcb0c1d7429797b2c5bc911aa2cf076e
Refactor guts of _find_interesting_networks

_find_interesting_networks is very sadly WAY too long and ugly. A
symptom of that was the code being all squished against the 80 column
line - but honestly the problem is that the code is so deep inside of
conditionals that figuring out what's going on is a bit madenning.

Split it out into another method so that the logic of whether we need to
do things can be isolated from the logic of doing things.

Change-Id: I4d5a30e3a5abd2b2f71e49dd18eab3eeb8f02f1d