Updated openstack/openstack Project: openstack-infra/shade  0c419aee9fa1aed5bd7ae92c0ff6654ba6a75e8b
Add nat_destination filter to floating IP creation

In Ansible 1.9.x 'quantum_floating_ip' module you have the option
to specify 'internal_network_name'. This allows you to set the name
of the network of the port to associate with the floating ip.

In Ansible 2.x 'os_floating_ip' you can't do that. This change
should bring the 'internal_network_name' option to be supported
for Ansible 2.x version.

Change-Id: I2c04a1feb5268a3621e35ed41a6af16622509325