Updated openstack/openstack Project: openstack/ironic-specs  a75428a0491059904de8599c507df5aa8faa84ee
Dynamic allocation of nodes on OneView drivers

This spec is about adding the ability to the OneView drivers of
dynamically allocate OneView resources to Ironic. The current
version of the drivers consider what we call "pre-allocation" of
nodes, meaning that when a node is registered in Ironic, even if
it is not in use, this resource is still reserved in OneView.
This change will prevent such situations by allocating OneView
resources only at boot time, allowing both systems to really
share the same pool of hardware.

Co-Authored-By: Sinval Vieira <sinval@lsd.ufcg.edu.br>
Co-Authored-By: Gabriel Bezerra <gabrielb@lsd.ufcg.edu.br>
Co-Authored-By: Thiago Paiva <thiagop@lsd.ufcg.edu.br>

Change-Id: Iace4d6cec1744b59dd2b15c59e1b7306295e88aa
Partial-Bug: #1541096