Updated openstack/openstack Project: openstack/requirements  6a08a63398491b52366f6336aa164289838ecae6
Downgrade docker-py from 1.8.0 to 1.7.2

docker-py 1.8.0 introduced a dependency py2-ipaddress, which is
imcompatible with another dependency cryptography [1]. This issue
broke Magnum [2] so we need to downgrade docker-py to remove
py2-ipaddress for now.

Note that this is a short term solution. In the long term, we will
work with upstream community to resolve the module conflicting
issue.

[1] https://github.com/pyca/cryptography/issues/2870
[2] http://lists.openstack.org/pipermail/openstack-dev/2016-April/
091916.html

Change-Id: I199c28017f8b18177b824e5594d09edd61798636
Related-Bug: #1568427