Updated openstack/openstack Project: openstack/keystone  a404ab4602a607b3a9fdd37b6aa7794f6713fc03
Set the values for the request_local_cache

Instead of only using the request_local_cache when an explicit set
occurs, make sure we always set the value for the request_local_cache
when we have to reach for the proxied backend.

The Context Local cache was being used previously when we were seeing
a mix of cache misses and hits. This change ensures we now always
set the value(s) as expected:

Change-Id: I4857cfe1e62d54c3c89a0206ffc895c4cf681ce5
Closes-Bug: #1567403