Updated openstack/openstack Project: openstack-infra/release-tools  04191338cc22d195a0331de7df3175b7a4a3df85
Pick last SHA for PREVIOUS_VERSION

sahara-tests-announce-release job failed. It was the first
release and we went looking for the first commit. Since
sahara-tests had git surgery when it was split from tempest,
we get 2 SHA's when we run "git rev-list --max-parents=0 HEAD".

So we should pick the last one.

Failure log:
http://logs.openstack.org/31/31a4756aa8ba3185a354dc8aa9bde02ded3c4357/release/sahara-tests-announce-release/a803ba5/console.html

Change-Id: I6a695e9d4248bafbde24cee95dc96a358fe285c1