Updated openstack/openstack Project: openstack-infra/project-config  24daa741da0c3ce554477cbecdf7d3237a589ff1
Rework slave-cleanup

The current approach does not work at all: Every job runs "git clean" as
part of set up of workspace. So, the marker we use is deleted and all
the content as well.

So, what we can do is run the "git clean" also after a run, so that the
workspace is minimal between invocations and the initial "git clean"
becomes a nop. This helps with overall space on the proposal slave
since only the active job will use .tox directories etc. Since there's
no regular cleanup of workspace, this also helps when jobs get retired,
the workspace will continue to exist - but with the cleanup done, it's
much smaller.

Change-Id: I8b17724905634aa2f7cc657de073ee7cfe10f781