Updated openstack/openstack Project: openstack/glance  52df790c66028f2261122c72b3fa55be2eb3f44f
Use messaging notifications transport instead of default

The usage of oslo_messaging.get_transport is not meant for
notifications; And oslo_messaging.get_notification_transport is
favored for those means. So this change introduces the usage of that
function.

If the settings for the notifications are not set with the
configuration that's under the oslo_messaging_notifications group,
this will fall back to the old settings which are under the DEFAULT
group; just like this used to work.

Closes-Bug: #1567231

Change-Id: I137d0f98533d8121334e8538d321fb6c7495f35f