Updated openstack/openstack Project: openstack/nova  19712dcccea548428cb3475e19aa492769deff2b
use 'required' instead of 'optional' for parameters

In swagger the attribute for required/optional is 'required', so we've
decided that's a better way to specify our parameters. This changes
the sphinx parser to use that when inserting (optional) into the
table.

We also make all 'path' parameters required, because we never specify
them in a table where they are not. I think them all being required:
false was an artifact of the data processing.

Part of bp:api-ref-in-rst

Change-Id: Ie2d012741d69430546fe1b10c9f16c853acb926a