Updated openstack/openstack Project: openstack/neutron  b98f2134b246e1337761be9a7846d61117f8ad8b
Add an option for WSGI pool size

Neutron currently hardcodes the number of
greenlets used to process requests in a process to 1000.
As detailed in
http://lists.openstack.org/pipermail/openstack-dev/2015-December/082717.html

this can cause requests to wait within one process
for available database connection while other processes
remain available.

By adding a wsgi_default_pool_size option functionally
identical to that of Nova, we can lower the number of
greenlets per process to be more in line with a typical
max database connection pool size.

DocImpact: a previously unused configuration value
           wsgi_default_pool_size is now used to affect
           the number of greenlets used by the server. The
           default number of greenlets also changes from 1000
           to 100.
Change-Id: I94cd2f9262e0f330cf006b40bb3c0071086e5d71