Updated openstack/openstack Project: openstack-dev/devstack  03b0f0b845cde1bc3e53e3034084e37eba510cd0
Remove driver certs directory

Cinder was the only project with a script in driver_certs. That cert is
no longer required for Cinder as we now just require that third party CI
is run for all drivers before acceptance.

This removes the cinder cert test script and the now empty driver_certs
directory.

Change-Id: I8d0867c4720f324b8dbf3c64ddf66ff267996d10