Updated openstack/openstack Project: openstack/ironic-specs  3c9a74b3a3e25d34283fd4d55d671d7ff68526a0
Update completed specs after Mitaka release

This update adds 5.0/ and 5.1/ directories, moves the symlinks for
completed specs into them, and updates the index.html page to reflect
the completion of the Mitaka cycle.

It also moves te partition-image-support-for-agent-driver from the
backlog to the approved directory, since the work was actually
completed, and updates the spec to include two missing (and unaffected)
sections.

Change-Id: Ice187410798a4d93044095dd476792d919222701