Updated openstack/openstack Project: openstack/glance  bf56e4e40c78b44595b8f781468bc93ae1d48a2f
Increase max wait time, avoid racy failure in gate

The test_task_api fails intermittently in gate.

Example:
http://logs.openstack.org/15/293415/2/check/gate-glance-python27/6a96731/testr_results.html.gz

The reason behind this is that the recent addition of flows to tasks
take extra time. When this test was written it considered execution time
for limited script length. Some of the nodes in the gate might be slower
than others resulting into these intermittent failures.

Until we come up with better way to control the flows that run by
default on tasks, we don't have a good way to run smaller script. The
increase of timeout doesn't hurt as much as the sleep interval is 0.05
seconds so, only the slower runs in the gate/dev box will be affected.

Co-Authored-By: Nikhil Komawar<nik.komawar@gmail.com>

Change-Id: I80f04f8b6a23c208f9553b398c8f5db05190bebb