Updated openstack/openstack Project: openstack/zaqar  3a2ab7160c916ab2bc542184d28c929f750cc539
expires should be checked when using pre-signed url

When creating pre-signed url, if expire time is set
as some special string like "'expires'='3600'",
after creation is done, the expire time will be
"3600-01-01T00:00:00", that means this url will be
valid until the year 3600!

So we should check the data format of expire time,
if it's a string of integer value, zaqar should return error.

Change-Id: I61d2d977d45b5d33ff3355f2c854215ee1a52bb2
Closes-Bug: #1561311