Updated openstack/openstack Project: openstack-infra/storyboard-webclient  fae86518e14a6f86f6e73985d52b1234c1666ba6
Make it clear when user reaches end of events timeline

Before there was no padding at the bottom, so a user would think
they could still scroll, when they could scroll no more. This adds
a horizontal rule to the end of the events timeline, so the user knows
that they have reached the end.

Change-Id: Icbc148a7c257067c804055fbb0a6aa7c89e33681