Updated openstack/openstack Project: openstack-infra/shade  74765c0e200ec3f9483fde1d5be81d60714517dc
Honor default_network for interface_ip

When we generate the interface_ip (which is used, amongst other things,
by ansible dynamic inventory) we do our best to figure out based on
config which of the IPs a server has should be the one that the user
wants to be "the" interface. In the previous patches, we added support
for the user configuring a "default_network" for one of their networks.
If the user has done this, and the server has an IP on that network, we
should use that network for interface_ip.

Change-Id: I440916622f9dfe12f8865bb1841dd0932d3cd7d0