Updated openstack/openstack Project: openstack-infra/tripleo-ci  29bff4153f2c289a1b681df6f1aa3d9f400fe8b4
Only use mirrored cache for git repos that exist

Adding repositories to mirrored.list makes CI try and use them, but
they wont exist until the patch has merged and everything is synced.
Allow for git repositories that are in the list but not actually on
the mirror server yet.

Change-Id: I8e9e0b90660d5791377b8bcc112b1ae068718ead