Updated openstack/openstack Project: openstack/nova  cdc899fdd50a43b633269d9ff76c2137fd2c2de2
claims: Do not assume image-meta is a dict

MoveClaim is expected to accept the image_meta argument to it's __init__
which we assumed was always a dict. This is not the case anymore as we
pass in the ImageMeta object, that recently lost it's dict-mimicing
powers so the assumption that it's safe to call from_dict on it is no
longer true.

This patch make sure that we convert to objects internally always.

Change-Id: I72081f656b02aa564af5461b24c5294b6bac59c8