Updated openstack/openstack Project: openstack-dev/pbr  6aa11493983739a88ab9de7af7cbbe68fde8dd03
Correct server test

This test was relying on the short sleep allowing the subprocess
time to write to an external file. Rather than write to an
external file, have the output go to a pipe that the test process
can read so that the test doesn't have to guess how long to sleep.

Change-Id: I128a9dc5c7525d941f6e00c0073d983eab1e44e6