Updated openstack/openstack Project: openstack-infra/storyboard-webclient  7af7433a7231bef88b22998770aefdb11db313ab
Make it possible to remove filters when creating a new worklist

At the moment, if you create a filter and then change your mind,
you have to close and reopen the worklist creation modal because
the delete button on the filter doesn't work.

This patch fixes this oversight by making the delete button on
the filter work.

Change-Id: Ifeb1ca814d194b4fddf78969151e9761e4efe2e2