Updated openstack/openstack Project: openstack-infra/zuul  33f080a8801dabac5450892b164ca510a8296405
Cloner: Don't fall back on infrastructure failure

Currently zuul-cloner falls back to the fallback branch when fetching
of the zuul ref fails. This is intended when the zuul ref is not found
on the remote. But in case the fetch fails due to infrastructure
reasons (e.g. zuul-merger is not reachable or certificate verification
failed) it should bail out with an error. Otherwise a already merged
and tested patch will be verified which could lead to broken patches
being merged.

Change-Id: Iefc82603de279e36ad5972ce341b102c8d38f69e