Updated openstack/openstack Project: openstack/diskimage-builder  4f6ce09385d567255649592bf5f93b7f3ef6eb04
Support to add certificate in ironic-agent

This commits provides support to add certificate while
building the image using ironic-agent element. The
certificate can be CA certificate or self-signed certificate.

The certificate is set to the environment variable
'DIB_IPA_CERT' which in turn is used by the ironic-agent
element while building the image.

Change-Id: I648f7934d4787dcc3030885cfca771b642a9595e