Updated openstack/openstack Project: openstack/neutron  4e3a9c2b9c0ada6c2a471d24a20927aa8e8f5740
Be explicit about the APIs tested in the gate

Add an rc hook that declaratively define what extensions
are going to be tested in the gate.

This has two benefits: a) it provides an easy switch to
disable intermittent/experimental APIs; b) it provides a
safeguard against accidental configuration errors that
may trigger the execution of dormant tests.

The 'service_type' extension has been skipped in preparation
of [1].

[1] https://review.openstack.org/#/c/195329/

Change-Id: I03e2394011aa747899b50e6eb64faca68866b388