Updated openstack/openstack Project: openstack-dev/devstack  406b45b81a55cb56b6695437dc8e1d6e5bd6e504
Remove shocco docs and other cleanups

This is a fairly opinionated change to do some spring cleaning on the
documentation.

The current output of shocco as rendered at [1] is completely broken.
I can not see that it is worth us maintaining this.  Honestly, the
github page does a better job at showing the scripts with a bit of
formatting.  The "changes" page is similarly useless today.  cgit or
github show allow browsing of changes in the repo better.  Both are
removed along with support scripts.

When you currently hit the first page, it gives no clue as to what
DevStack actually is.  Add a paragraph explaining that, and link to
the cgit for easy source browsing.

stackrc.rst is not necessary; the stuff about database backends is
already discussed in configuration.rst; move the things about service
repos into a section of configuration.rst.

The discussion in openrc.rst is moved into the configuration.rst file.

localrc.conf.rst was just a paragraph pointing back to
configuration.rst; this is removed.

The variables described in exercise.rst are moved into a separate
section of configuration.rst

[1] http://docs.openstack.org/developer/devstack/#scripts

Change-Id: Ie7f4b265368f1d10a8908d75e11d625b2cc39e7c