Updated openstack/openstack Project: openstack/python-heatclient  4763d5d8fc05247b10a8c2757493278c588fd20a
Moved -t parameter for visibility

When called without the required arguments, the allowed arguments are
output to the screen. The order of those arguments is dictated by when
they are added to the parser.

Before this change, -t (a required parameter) was at the start of the
list. This obfuscated the required parameter list when the help output
was displayed:

usage: openstack stack create [-h]
                              [-f
{html,json,json,shell,table,value,yaml,yaml}]
                              [-c COLUMN] [--max-width <integer>]
[--noindent]
                              [--prefix PREFIX] -t <template>
                              [-e <environment>] [--timeout <timeout>]
                              [--pre-create <resource>]
[--enable-rollback]
                              [--parameter <key=value>]
                              [--parameter-file <key=file>] [--wait]
                              [--tags <tag1,tag2...>] [--dry-run]
                              <stack-name>

Notice how the -t appears in the middle of the optional arguments. This
change moves it to the end so that it is more easily visible to the user
what is required:

usage: openstack stack create [-h]
                              [-f
{html,json,json,shell,table,value,yaml,yaml}]
                              [-c COLUMN] [--max-width <integer>]
[--noindent]
                              [--prefix PREFIX] [-e <environment>]
                              [--timeout <timeout>] [--pre-create
<resource>]
                              [--enable-rollback] [--parameter
<key=value>]
                              [--parameter-file <key=file>] [--wait]
                              [--tags <tag1,tag2...>] [--dry-run]
                              -t <template>
                              <stack-name>

Change-Id: I4e17650622dc8e206dc1aa39e4102f16429ac353