Updated openstack/openstack Project: openstack-infra/project-config  cf6e9fca3dc24786977684faaeef1283964c8e6c
Switch gate-neutron-dsvm-api to be invoked via tempest plugin framework

Neutron was amongst one of the earliest projects that promoted the
idea of colocating more advanced tests within the project's source
repo itself. The job gate-neutron-dsvm-api was created out of the
tempest backbone when no plugin/tempest-lib were a thing. Things
have matured since, therefore it is time to update some of the
plumbing underpinning this job so that we can start using some of
the latest cool stuff that Tempest has to offer.

The switch must be made conditional to branches since the it only
makes sense from Newton onward.

Depends-on: Ia233aa162746845f6ae08a8157dcd242dcd58eab
Change-Id: If5f7ea82c45494953452f728fabb7403d0b6e14e