Updated openstack/openstack Project: openstack-infra/tripleo-ci  8fa3717b8084393f8ccac5f5d9e9974a9e3bd992
Removing previously created resources in pingtest and --skip-pingtest-cleanup option

If the pingtest fails in somepoint, images and network resources
are still created, in the case of pingtest images, they are created
with different ID's and when referencing the i.e. pingtest_image
by name, an error is returned as there are more than one image with the same
name.

This patch in the cleanup, removes all pingtest_* images.
And at the beginning of the overcloud_pingtest method
removes all the previous images or the nova network.

Also for debuging purposes a --skip-pingtest-cleanup option is added,
which allows not to delete the resources create in the pingtest

Change-Id: I10670af03510a7735433e37009e37c59fd0af3f4
Closes-Bug: #1569309