Updated openstack/openstack Project: openstack-infra/system-config  8832f7c55efcbcb032a82a4882956c6ff959212b
Moved npm rewrite rules out of directory scope

mod_rewrite has inconsistent behavior inside of a directory tag. For
example, it doesn't appear to update the value of REQUEST_URI in
subsequent redirects, and the modifiable URI passed to the RewriteRule
is different from that passed to the RewriteCond.

This patch moves the rewrite rules for the npm mirror out of the
directory block, in order to fix unexpected 500 errors when accessing
packages with single file names. It also switches the <directory>
configuration block to only match on subdirectories of /npm/, so that
listing the contents of /npm does not redirect the user to the
/npm/index.json/ package.

Why that package exists is left as an exercise to the user.

Change-Id: Idbad41867d3d3546a59fe0c4fa54c5be22e31438