Updated openstack/openstack Project: openstack/nova  0b9f7538d3e47ac5afa357ab0c7647b6f020772a
Add instance/instance_uuid to build_requests table

New fields keep needing to be added to build_requests in order to
capture all of the relevant instance information necessary. At this
point it's simpler to just embed a serialized Instance object in the db.
A column for that is added, as well as an instance_uuid column for
indexed lookups.

Change-Id: I1628ddd8cb75b4d0a7bbcd2720536c2e0bfc6043
Partially-implements: bp add-buildrequest-obj