Updated openstack/openstack Project: openstack/nova  c41230be037be9289fd01915ddbf289500f726ef
Collection of CSS fixes

This ensures all parameter tables are 100% wide, even if they have
only a few items with short descriptions. It makes all the tables the
same width.

This moves the link icon into the gutter for the main content, and
makes it hidden by default (only visible on hover) like in the API
site today.

This makes the link + method name div be ``white-space: nowrap`` to
ensure that when the screen gets narrower the Delete button doesn't
fold up on itself.

Part of bp:api-ref-in-rst

Change-Id: I3892e6c8e3c01956d63a33603ba1bf9146d21161