Updated openstack/openstack Project: openstack/nova  e95e4cf023252e68d7a982ae87170dbc78911788
clean up versions.inc reference document

This fixes a number of issues in the versions.inc reference
document.

Documents all the useful parameters in the response (we really need to
do something about the ones that are not).

Fix links to samples content, and write some text explaining them.

Parametrize the {api_version} GET call, it was weird to list all the
hard coded values there.

Add a couple of inline todos for longer additional references we
should do.

Part of bp:api-ref-in-rst

Change-Id: I56e5e922fca39453d0d9fc062a84daba7cad8cc1