Updated openstack/openstack Project: openstack/nova  655922ca6647a95a0c5859bacc143198f5a66393
Add checks for driver attach_interfaces capability

Currently Ironic driver does not support attach interfaces
for instances, and the error of NotImplementedError has
not been handled.

Accodring to the comments of the previous patchset 3,
a better solution of this problem will be add a check
of driver capability before the actual interface attach
job, as it won't have the problem of orphan port [1]

This patch addes checks for driver attach interface
capabalities.

[1] https://review.openstack.org/#/c/272471/3/nova/compute/manager.py

Change-Id: I3fa376f866e70e1b90a6aa10b1fe6b169a0ffa43
Closes-Bug: #1538054