Updated openstack/openstack Project: openstack/python-neutronclient  7be2ca5b211d00b1774dd53ad64c212864fd078c
Devref: Newton updates for transition to OSC

Update the "Transition to OpenStack Client" devref with the plan
for handling network resources in Newton. In particular, this
update clarifies the OpenStack project where the commands will
reside. This update attempts to capture some of the information
in [1] and to address packaging concerns.

[1] https://etherpad.openstack.org/p/osc-neutron-support

Change-Id: I95e579ab23377a6cf36fc7a50356ab406d350053
Related-Bug: #1521291