Updated openstack/openstack Project: openstack-infra/project-config  79c67f045837cd92449cdba30d600346631878cb
Add experimental ubuntu-xenial jobs for OpenStack-Ansible roles

This patch adds experimental Ubuntu Xenial jobs to all the
OpenStack-Ansible roles. The jobs are '-nv' jobs so that if
the 'ansible-role-functional-jobs-xenial' is switched from
the experimental queue to the check/gate queue then these jobs
are not affected. It keeps things more flexible.

Change-Id: I23a5dafed406436ecf4299ae269c6eabe2fa07a5